Revert "V0.1.7 - Utils assert and fs methods. Better import (one object)"
This reverts commit 44548d2458e73e6d748868c25bfaad7e78d3c316.
