+++ CHANGELOG.md
@@ -15,19 +15,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [FEATURE] Work flow templates 
 - See up.todo for a broad overview of MVP (v1.0.0) features. 
 
-## [0.1.7] - 2018-11-11
+## [0.1.6] - 2018-11-11
 ### Added 
-
-
-## [0.1.6] - 2018-11-11 - [7d55d0e65987690c69457409780cfef8d732bbd8]
-### Added 
-- [FEATURE] Install.sh now installs Node.js (if not installed)
-- [FEATURE] Tree-model package
-
-### Changed 
 - [CHORE] build.sh moved to ```./scripts``` 
 - [CHORE] install.sh moved to ```./scripts``` 
 - [CHORE] correct link pointing to ```./srcipts/build.sh``` for ```build``` command.
+- [FEATURE] Install.sh now installs Node.js (if not installed)
+- [FEATURE] Tree-model package
 
 ## [0.1.5] - 2018-11-01 - [1c3454237b9862c79ad253adbc229b496c442ca4]
 ### Fixed
