+++ lib/utils/functional.js
@@ -5,6 +5,9 @@
  */
 'use strict'
 
+const fs = require('fs')
+const path = require('path')
+const junk = require('junk')
 const _ = require('lodash/fp')
 
 /**
@@ -34,24 +37,65 @@ const filter = _.curry((fn, xs) => xs.filter(fn))
 const store = _.memoize(_.values)
 
 /**
- * Loop over each key-value pair, preforming mutations to
- * each with a callback function
- * @param  {Object} object
- * @param  {Function} cbV    - callback for the object values
- * @param  {Function} cbK    - callback for the object keys
- * @return {Object}          - New object with mapped key/values
+ * Read the content of a directory
+ * @param  {String | Buffer} dir - path to dir
+ * @return {Array} - Names of items in dir
  */
-const mapObjectKeyValue = (object, cbV, cbK) =>
-  _.mapKeys(_.mapValues(object, v => cbV(v)), (v, k) => cbK(k))
+const readDir = dir => fs.readdirSync(dir)
 
 /**
- * Loop over each key-value pair, preforming a mutation
- * to each pair.
- * @param  {Object}   object
- * @param  {Function} cb     - callback that preforms a mutation
- * @return {Object}          - A mutated clone of the original object
+ * [description]
+ * @param  {[type]} path [description]
+ * @return {[type]}      [description]
  */
-const mapObject = (object, cb) => Object.entries(object).map(([key, value]) => cb(key, value))
+const requireFile = async path => await require(path)
+
+/**
+ * filterJunk :: String -> [String]
+ */
+const filterJunk = filter(junk.not)
+
+/**
+ * readDirAndFilter :: (a -> [String]) -> (b -> Bool) -> [String]
+ */
+const readDirAndFilter = compose(
+  filterJunk,
+  readDir
+)
+
+/**
+ * Resolve the path to a directory, based on the global __rootDir variable.
+ * @param  {String} location - path to the directory
+ * @return {String}          - resolved rootpath and directory location
+ */
+const resolveRoot = location => path.resolve(__rootDir, location)
+
+/**
+ * Resolve a path and require the file.
+ */
+const resolveAndRequire = compose(
+  resolveRoot,
+  require
+)
+
+/**
+ * Creates a directory if it does not exist. Handles some
+ * edge cases that could result in error while trying to
+ * determine whether a directory exists or nor and create that
+ * directory at the same time.
+ *
+ * @credit
+ * based on: https://stackoverflow.com/a/21196961 (by josh3736).
+ *
+ * @param  {string}   path - where the directory should be created.
+ * @param  {Function} cb   - Callback function, returns an error object.
+ * @return {Function}      - Callback function
+ */
+const dirEnsureExists = (path, cb) => {
+  fs.mkdir(path, err => {
+    !err ? cb(null) : err.code === 'EEXIST' ? cb(null) : cb(err)
+  })
+}
 
 /**
  * Exports
@@ -62,6 +106,8 @@ module.exports = {
   match,
   filter,
   store,
-  mapObjectKeyValue,
-  mapObject
+  readDir,
+  requireFile,
+  readDirAndFilter,
+  dirEnsureExists
 }
