+++ lib/utils/index.js
@@ -1,13 +1,3 @@
-const log = require('./log')
-const functional = require('./functional')
-const assert = require('./assert')
-const files = require('./files')
-const misc = require('./misc')
-
-module.exports = {
-    ...log,
-    ...functional,
-    ...assert,
-    ...files,
-    ...misc
-}
+module.exports.f = require('./functional')
+module.exports.log = require('./log')
+module.exports.misc = require('./misc')
