+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "up",
-  "version": "0.1.7",
+  "version": "0.1.6",
   "description": "A modular command line framework for automating tasks and common workflows",
   "author": "Yaro Schiffelers",
   "license": "MIT",
